% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/literanger-package.R
\docType{package}
\name{literanger-package}
\alias{literanger-package}
\title{\pkg{literanger}: Random forests for multiple imputation algorithms.}
\description{
'literanger' is an adaption of the 'ranger' R package for training and
predicting from random forest models within multiple imputation problems.
'ranger' is a fast implementation of random forests (Breiman, 2001) or
recursive partitioning, particularly suited for high dimensional data
(Wright et al, 2017a). literanger enables random forests to be embedded in
the fully conditional specification framework for multiple imputation known
as  'Multiple Imputation via Chained Equations' (Van Buuren 2007).
Specifically, literanger records terminal nodes' in-bag data as needed by
the algorithm described in Doove et al (2014). Alternatively, the usual
bagged prediction can be provided as in the imputation algorithm called
'missForest' (Stekhoven et al, 2014).
}
\details{
Classification and regression forests are implemented as in the original
Random Forest (Breiman, 2001) or using extremely randomized trees (Geurts et
al, 2006). "data.frame", "matrix", and sparse matrices ("dgCmatrix") are
supported.

Split selection may be based on improvement in metrics such as the variance,
Gini impurity, beta log-likelihood (Weinhold et al, 2019), Hellinger distance
(Cieslak et al, 2012) or maximally selected statistics (Wright et al, 2017b).

See \url{https://github.com/imbs-hl/ranger} for the development version of ranger
or \url{https://github.com/stephematician/literanger} for development version of
this package.

For alternative approaches to multiple imputation that employ random forests,
see 'missRanger' \url{https://cran.r-project.org/package=missRanger} and
'miceRanger' \url{https://cran.r-project.org/package=miceRanger}, which use
predictive mean matching combined with the original 'ranger' algorithm.

This package was adapted from the "ranger" package for R Statistical
Software. The C++ core is provided under the same license terms as the
original C++ core in the 'ranger' package, namely the MIT license
\url{https://www.r-project.org/Licenses/MIT}. The wrappers in this package around
the core are licensed under the same terms of the corresponding components in
the 'ranger' R package, namely the GPL3 license
\url{https://www.r-project.org/Licenses/GPL-3}, \url{http:://www.gnu.org/licenses}.

License statement for C++ core of ranger:

Copyright (c) \link{2014-2018} \link{Marvin N. Wright}

This software may be modified and distributed under the terms of the MIT
license.

Please note that the C++ core of ranger is distributed under MIT license
and the R package "ranger" under GPL3 license.

License statement for the additions to the core in the 'ranger' R package:

Ranger is free software: you can redistribute it and/or modify it under
the terms of the GNU General Public License as published by the Free
Software Foundation, either version 3 of the License, or (at your option)
any later version.

Ranger is distributed in the hope that it will be useful, but WITHOUT ANY
WARRANTY; without even the implied warranty of MERCHANTABILITY or FITNESS
FOR A PARTICULAR PURPOSE. See the GNU General Public License for more
details.

You should have received a copy of the GNU General Public License along
with Ranger. If not, see \url{http://www.gnu.org/licenses/}.

Written by:
Marvin N. Wright
Institut f체r Medizinische Biometrie und Statistik
Universit채t zu L체beck
Ratzeburger Allee 160
23562 L체beck
http://www.imbs-luebeck.d
}
\references{
\itemize{
\item Cieslak, D. A., Hoens, T. R., Chawla, N. V., & Kegelmeyer, W. P. (2012).
Hellinger distance decision trees are robust and skew-insensitive. \emph{Data
Mining and Knowledge Discovery}, 24, 136-158.
\doi{10.1007/s10618-011-0222-1}.
\item Doove, L. L., Van Buuren, S., & Dusseldorp, E. (2014). Recursive
partitioning for missing data imputation in the presence of interaction
effects. \emph{Computational Statistics & Data Analysis}, 72, 92-104.
\doi{10.1016/j.csda.2013.10.025}.
\item Geurts, P., Ernst, D., & Wehenkel, L. (2006). Extremely randomized trees.
\emph{Machine Learning}, 63, 3-42. \doi{10.1007/s10994-006-6226-1}.
\item Stekhoven, D.J. and Buehlmann, P. (2012). MissForest--non-parametric
missing value imputation for mixed-type data. \emph{Bioinformatics}, 28(1),
112-118. \doi{10.1093/bioinformatics/btr597}.
\item Van Buuren, S. (2007). Multiple imputation of discrete and continuous
data by fully conditional specification. _Statistical Methods in Medical
Research, 16(3), 219-242. \doi{10.1177/0962280206074463}.
\item Weinhold, L., Schmid, M., Wright, M. N., & Berger, M. (2019). A random
forest approach for modeling bounded outcomes. \emph{arXiv preprint},
arXiv:1901.06211. \doi{10.48550/arXiv.1901.06211}.
\item Wright, M. N., & Ziegler, A. (2017a). ranger: A Fast Implementation of
Random Forests for High Dimensional Data in C++ and R. \emph{Journal of
Statistical Software}, 77, 1-17. \doi{10.18637/jss.v077.i01}.
\item Wright, M. N., Dankowski, T., & Ziegler, A. (2017b). Unbiased split
variable selection for random survival forests using maximally selected
rank statistics. \emph{Statistics in medicine}, 36(8), 1272-1284.
\doi{10.1002/sim.7212}.
}
}
\author{
Stephen Wade \href{mailto:stephematician@gmail.com}{stephematician@gmail.com}, Marvin N Wright (original
'ranger' package)
}
