% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/predict.R
\name{predict.literanger}
\alias{predict.literanger}
\title{Ranger prediction}
\usage{
\method{predict}{literanger}(
  object,
  newdata = NULL,
  prediction_type = c("bagged", "doove"),
  seed = sample.int(n = .Machine$integer.max, size = 1),
  n_thread = 0,
  verbose = F,
  ...
)
}
\arguments{
\item{object}{A trained random forest \code{literanger} object.}

\item{newdata}{New test data of class \code{data.frame} or `dgCmatrix``.}

\item{seed}{Random seed. Default is \code{NULL}, which generates the seed from
\code{R}. Set to \code{0} to ignore the \code{R} seed.}

\item{verbose}{Verbose output on or off.}

\item{num.threads}{Number of threads. Default is number of CPUs available.}
}
\value{
Object of class \code{literanger.prediction} with elements:
\describe{
\item{\code{predictions}}{Predicted (drawn) classes/value for classification and
regression.}
\item{\code{num.trees}}{Number of trees.}
\item{\code{num.independent.variables}}{Number of independent variables.}
\item{\code{treetype}}{Type of forest/tree. Classification, regression or
survival.}
\item{\code{num.samples}}{Number of samples.}
}
}
\description{
Predict with new data and a ranger object from Ranger.
}
\details{
The predicted class (classification) or numeric value (regression) is
returned by drawing a tree and then a sample from the tree for each row of
the data.

Factor levels are returned as numerics. To retrieve the corresponding factor
levels, use \code{rf$forest$levels}, if \code{rf} is the ranger object.
}
\examples{
## Classification forest
train.idx <- sample(nrow(iris), 2/3 * nrow(iris))
iris.train <- iris[train.idx, ]
iris.test <- iris[-train.idx, ]
rg.iris <- train_lite(data=iris.train, response_name="Species")
pred.iris <- predict(rg.iris, newdata=iris.test)
table(iris.test$Species, pred.iris$values)

}
\references{
\itemize{
\item Wright, M. N. & Ziegler, A. (2017). ranger: A Fast Implementation of
Random Forests for High Dimensional Data in C++ and R. J Stat Softw
77:1-17. \doi{10.18637/jss.v077.i01}.
}
}
\seealso{
\code{\link{train}}
}
\author{
Marvin N. Wright
}
